refactor: spell changes in name file
