+++ src/reducer/functionsNews.js
